server/requirements.txt
pillow
google-generativeai
flask
flask-cors
python-dotenv
mysql-connector-python
watchdog
Faker